+++ lib/src/typedefs/on_tab_remove.dart
@@ -1,4 +1,0 @@
-import '../tab_data.dart';
-
-/// The event triggered when the tab is removed.
-typedef OnTabRemove = void Function(TabData tabData);